StudentManagement System
The main aim to design a’ student management system’. STUDENT list to the teacher. Using spring boot to create project output. Eclipse workspace without installing to local maven repository. Hibernate to provides a framework for mapping an object-oriented domain model to relational database. It generates SQL calls and relieves the developer from the manual handling and object conversion of the result set.